Re #2407 Move EnvConfig out of Stack.Types.Config
