+++ package.yaml
@@ -267,6 +267,7 @@ library:
   - Stack.Types.CabalConfigKey
   - Stack.Types.ColorWhen
   - Stack.Types.CompilerBuild
+  - Stack.Types.CompilerPaths
   - Stack.Types.Compiler
   - Stack.Types.Config
   - Stack.Types.Config.Build
@@ -275,6 +276,8 @@ library:
   - Stack.Types.DockerEntrypoint
   - Stack.Types.DownloadInfo
   - Stack.Types.DumpLogs
+  - Stack.Types.DumpPackage
+  - Stack.Types.EnvConfig
   - Stack.Types.GHCDownloadInfo
   - Stack.Types.GHCVariant
   - Stack.Types.GhcOptionKey
@@ -287,6 +290,7 @@ library:
   - Stack.Types.Nix
   - Stack.Types.Package
   - Stack.Types.PackageName
+  - Stack.Types.PlatformVariant
   - Stack.Types.PvpBounds
   - Stack.Types.Resolver
   - Stack.Types.Runner
