+++ src/Stack/Runners.hs
@@ -35,9 +35,10 @@ import           Stack.Setup ( setupEnv )
 import           Stack.Storage.User ( logUpgradeCheck, upgradeChecksSince )
 import           Stack.Types.ColorWhen ( ColorWhen (..) )
 import           Stack.Types.Config
-                   ( BuildOptsCLI, Config (..), EnvConfig, defaultBuildOptsCLI )
+                   ( BuildOptsCLI, Config (..), defaultBuildOptsCLI )
 import           Stack.Types.ConfigMonoid ( ConfigMonoid (..) )
 import           Stack.Types.Docker ( dockerEnable )
+import           Stack.Types.EnvConfig ( EnvConfig )
 import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
 import           Stack.Types.Nix ( nixEnable )
 import           Stack.Types.Runner
