+++ src/Stack/Storage/User.hs
@@ -49,10 +49,10 @@ import           Stack.Types.Build ( PrecompiledCache (..) )
 import           Stack.Types.Cache ( Action (..) )
 import           Stack.Types.Compiler ( ActualCompiler, compilerVersionText )
 import           Stack.Types.CompilerBuild ( CompilerBuild )
+import           Stack.Types.CompilerPaths
+                   ( CompilerPaths (..), GhcPkgExe (..) )
 import           Stack.Types.Config
-                   ( CompilerPaths (..), GhcPkgExe (..), HasConfig (..)
-                   , UserStorage (..), configUserStorage
-                   )
+                   ( HasConfig (..), UserStorage (..), configUserStorage )
 import           System.Posix.Types ( COff (..) )
 import           System.PosixCompat.Files
                    ( fileSize, getFileStatus, modificationTime )
